refactor: dropdown 그룹화 (#17)
